Refactor: Restructure project and improve parser error handling
This commit includes the following changes:

- Restructured the project into a more standard Go layout with `cmd` and `pkg` directories.
- Updated all import paths to reflect the new structure.
- Initialized a Go module with `go mod init`.
- Improved the parser's error handling to recover from errors and continue parsing.
- Modified the main function to print parsing errors.

I was stuck on the error handling part, where the program was terminating without any output. I realized that the panic was being caught, but the errors were not being printed. I then fixed this by modifying the main function to print the errors.
