+++ go.mod
@@ -1,0 +1,3 @@
+module gopy
+
+go 1.24.3
