+++ pkg/lexer/lexer_test.go
