+++ test.gopy
@@ -1,0 +1,3 @@
+let x = 5
+let y = 10 + ;
+let z = 20
